Added Mp3Track for easy processing of MP3 tracks
