+++ src/main/java/ml/duncte123/skybot/objects/audiomanagers/speech/SpeechAudioTrack.java
@@ -18,55 +18,20 @@
 
 package ml.duncte123.skybot.objects.audiomanagers.speech;
 
-import com.sedmelluq.discord.lavaplayer.container.mp3.Mp3AudioTrack;
-import com.sedmelluq.discord.lavaplayer.source.AudioSourceManager;
-import com.sedmelluq.discord.lavaplayer.tools.io.HttpInterface;
-import com.sedmelluq.discord.lavaplayer.tools.io.PersistentHttpStream;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrackInfo;
-import com.sedmelluq.discord.lavaplayer.track.DelegatedAudioTrack;
-import com.sedmelluq.discord.lavaplayer.track.playback.LocalAudioTrackExecutor;
 import ml.duncte123.skybot.Author;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.net.URI;
+import ml.duncte123.skybot.objects.audiomanagers.Mp3Track;
 
 @Author(nickname = "ramidzkh", author = "Ramid Khan")
-public class SpeechAudioTrack extends DelegatedAudioTrack {
-
-    private static final Logger log = LoggerFactory.getLogger(SpeechAudioTrack.class);
-
-    private final SpeechAudioSourceManager manager;
+public class SpeechAudioTrack extends Mp3Track {
 
     SpeechAudioTrack(AudioTrackInfo trackInfo, SpeechAudioSourceManager manager) {
-        super(trackInfo);
-        this.manager = manager;
-    }
-
-    @Override
-    public void process(LocalAudioTrackExecutor executor) throws Exception {
-        try (HttpInterface httpInterface = manager.getHttpInterface()) {
-            loadStream(executor, httpInterface);
-        }
-    }
-
-    private void loadStream(LocalAudioTrackExecutor localExecutor, HttpInterface httpInterface) throws Exception {
-        String trackUrl = trackInfo.identifier;
-        log.debug("Starting Speech speech from URL: {}", trackUrl);
-
-        try (PersistentHttpStream stream = new PersistentHttpStream(httpInterface, new URI(trackUrl), null)) {
-            processDelegate(new Mp3AudioTrack(trackInfo, stream), localExecutor);
-        }
+        super(trackInfo, manager);
     }
 
     @Override
     public AudioTrack makeClone() {
-        return new SpeechAudioTrack(trackInfo, manager);
-    }
-
-    @Override
-    public AudioSourceManager getSourceManager() {
-        return manager;
+        return new SpeechAudioTrack(trackInfo, (SpeechAudioSourceManager) getSourceManager());
     }
 }
